Make rollingrestart a method on statefulsets
