+++ cluster-controller/src/main/java/io/strimzi/controller/cluster/operations/resource/ResourceOperation.java
@@ -116,7 +116,7 @@ public abstract class ResourceOperation<C, T extends HasMetadata, L, D, R extend
         );
     }
 
-    public <P extends Patchable<T, T>> void patch(String namespace, String name, T patch, Handler<AsyncResult<Void>> handler) {
+    public void patch(String namespace, String name, T patch, Handler<AsyncResult<Void>> handler) {
         patch(namespace, name, true, patch, handler);
     }
 
@@ -124,7 +124,7 @@ public abstract class ResourceOperation<C, T extends HasMetadata, L, D, R extend
         return operation().inNamespace(namespace).withName(name).get();
     }
 
-    public <P extends Patchable<T, T>> void patch(String namespace, String name, boolean cascading, T patch, Handler<AsyncResult<Void>> handler) {
+    public void patch(String namespace, String name, boolean cascading, T patch, Handler<AsyncResult<Void>> handler) {
         vertx.createSharedWorkerExecutor("kubernetes-ops-pool").executeBlocking(
                 future -> {
                     try {
